+++ src/population.cpp
@@ -71,6 +71,6 @@ void Population::resize(size_t s){
     _individuals.resize(s);
     _indexes.resize(s);
     int counter = 0;
-    for(int &i: _indexes)
+    for(size_t &i: _indexes)
         i = counter++;
 }
